IGNITE-25208 Make BinaryObjectBuilderImpl package private
