+++ modules/calcite/src/main/java/org/apache/ignite/internal/processors/query/calcite/schema/CacheTableDescriptorImpl.java
@@ -45,7 +45,7 @@ import org.apache.ignite.binary.BinaryObject;
 import org.apache.ignite.binary.BinaryObjectBuilder;
 import org.apache.ignite.cache.CacheWriteSynchronizationMode;
 import org.apache.ignite.cluster.ClusterNode;
-import org.apache.ignite.internal.binary.BinaryUtils;
+import org.apache.ignite.internal.binary.builder.BinaryObjectBuilders;
 import org.apache.ignite.internal.processors.affinity.AffinityTopologyVersion;
 import org.apache.ignite.internal.processors.cache.CacheStoppedException;
 import org.apache.ignite.internal.processors.cache.GridCacheContext;
@@ -406,7 +406,7 @@ public class CacheTableDescriptorImpl extends NullInitializerExpressionFactory
 
         BinaryObjectBuilder builder = cctx.grid().binary().builder(typeName);
 
-        BinaryUtils.prepareAffinityField(builder, cctx.cacheObjectContext());
+        BinaryObjectBuilders.prepareAffinityField(builder, cctx.cacheObjectContext());
 
         return builder;
     }
@@ -477,7 +477,7 @@ public class CacheTableDescriptorImpl extends NullInitializerExpressionFactory
         BinaryObjectBuilder builder = cctx.grid().binary().builder(
             cctx.grid().binary().<BinaryObject>toBinary(val));
 
-        BinaryUtils.prepareAffinityField(builder, cctx.cacheObjectContext());
+        BinaryObjectBuilders.prepareAffinityField(builder, cctx.cacheObjectContext());
 
         return builder;
     }
