+++ modules/core/src/main/java/org/apache/ignite/internal/client/thin/ClientBinary.java
@@ -29,6 +29,7 @@ import org.apache.ignite.internal.binary.BinaryContext;
 import org.apache.ignite.internal.binary.BinaryEnumObjectImpl;
 import org.apache.ignite.internal.binary.BinaryMetadata;
 import org.apache.ignite.internal.binary.BinaryUtils;
+import org.apache.ignite.internal.binary.builder.BinaryObjectBuilders;
 import org.apache.ignite.internal.binary.streams.BinaryStreams;
 
 /**
@@ -68,7 +69,7 @@ public class ClientBinary implements IgniteBinary {
         if (typeName == null || typeName.isEmpty())
             throw new IllegalArgumentException("typeName");
 
-        return BinaryUtils.createBuilder(binaryContext(), typeName);
+        return BinaryObjectBuilders.createBuilder(binaryContext(), typeName);
     }
 
     /** {@inheritDoc} */
@@ -76,7 +77,7 @@ public class ClientBinary implements IgniteBinary {
         if (binaryObj == null)
             throw new NullPointerException("binaryObj");
 
-        return BinaryUtils.toBuilder(binaryObj);
+        return BinaryObjectBuilders.toBuilder(binaryObj);
     }
 
     /** {@inheritDoc} */
