+++ modules/core/src/test/java/org/apache/ignite/internal/binary/BinaryMarshallerSelfTest.java
@@ -80,6 +80,7 @@ import org.apache.ignite.binary.BinaryWriter;
 import org.apache.ignite.binary.Binarylizable;
 import org.apache.ignite.configuration.BinaryConfiguration;
 import org.apache.ignite.configuration.IgniteConfiguration;
+import org.apache.ignite.internal.binary.builder.BinaryObjectBuilders;
 import org.apache.ignite.internal.binary.streams.BinaryInputStream;
 import org.apache.ignite.internal.binary.streams.BinaryOutputStream;
 import org.apache.ignite.internal.binary.streams.BinaryStreams;
@@ -2502,7 +2503,7 @@ public class BinaryMarshallerSelfTest extends AbstractBinaryArraysTest {
      * @return Copy.
      */
     private BinaryObject copy(BinaryObject po, Map<String, Object> fields) {
-        BinaryObjectBuilder builder = BinaryUtils.toBuilder(po);
+        BinaryObjectBuilder builder = BinaryObjectBuilders.toBuilder(po);
 
         if (fields != null) {
             for (Map.Entry<String, Object> e : fields.entrySet())
@@ -3041,7 +3042,7 @@ public class BinaryMarshallerSelfTest extends AbstractBinaryArraysTest {
         marsh = binaryMarshaller();
 
         // Checking the builder.
-        BinaryObjectBuilder builder = BinaryUtils.createBuilder(binaryContext(marsh),
+        BinaryObjectBuilder builder = BinaryObjectBuilders.createBuilder(binaryContext(marsh),
             "org.gridgain.foo.bar.TestClass");
 
         builder.setField("a", "1");
@@ -3725,7 +3726,7 @@ public class BinaryMarshallerSelfTest extends AbstractBinaryArraysTest {
 
         BinaryMarshaller m = binaryMarshaller();
 
-        BinaryObjectBuilder builder = BinaryUtils.createBuilder(binaryContext(m), "MyFakeClass");
+        BinaryObjectBuilder builder = BinaryObjectBuilders.createBuilder(binaryContext(m), "MyFakeClass");
 
         String[] fieldNames = {"field9", "field8", "field0", "field1", "field2"};
 
@@ -4244,7 +4245,7 @@ public class BinaryMarshallerSelfTest extends AbstractBinaryArraysTest {
 
         marsh.setBinaryContext(ctx, iCfg);
 
-        return BinaryUtils.createBuilder(ctx, typeName);
+        return BinaryObjectBuilders.createBuilder(ctx, typeName);
     }
 
     /**
