+++ modules/core/src/test/java/org/apache/ignite/testframework/junits/IgniteMock.java
@@ -65,7 +65,7 @@ import org.apache.ignite.internal.IgniteEx;
 import org.apache.ignite.internal.binary.BinaryCachingMetadataHandler;
 import org.apache.ignite.internal.binary.BinaryContext;
 import org.apache.ignite.internal.binary.BinaryMarshaller;
-import org.apache.ignite.internal.binary.BinaryUtils;
+import org.apache.ignite.internal.binary.builder.BinaryObjectBuilders;
 import org.apache.ignite.internal.cluster.IgniteClusterEx;
 import org.apache.ignite.internal.management.IgniteCommandRegistry;
 import org.apache.ignite.internal.processors.cache.GridCacheUtilityKey;
@@ -459,7 +459,7 @@ public class IgniteMock implements IgniteEx {
 
             /** {@inheritDoc} */
             @Override public BinaryObjectBuilder builder(String typeName) throws BinaryObjectException {
-                return BinaryUtils.createBuilder(ctx, typeName);
+                return BinaryObjectBuilders.createBuilder(ctx, typeName);
             }
         };
